+++ lib/is_external_link.js
